// src/app/models/chamados.ts
export interface Chamado {
    id?:                any;
    dataAbertura?:      string;
    dataFechamento?:    string;
    status:             string;
    titulo:             string;
    observacoes:          string;
    cartao:             any;
    cliente:            any;
    nomeCliente:        string;
    nomeCartao:         string; 
}